修改连接超时、读超时参数类型

+++ op-framework/op-boot/op-boot-mall-starter/src/main/java/com/op/mall/client/jingdong/JdMallProperties.java
@@ -17,12 +17,12 @@ public class JdMallProperties {
     /**
      * 连接超时
      */
-    private int connectTimeout;
+    private long connectTimeout;
 
     /**
      * 读超时
      */
-    private int readTimeout;
+    private long readTimeout;
 
     /**
      * 待明确
